# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Daily Management Express API",
    "description": "Daily Management REST API documentation for mobile client and server integration.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3005",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Enter the Better Auth session token"
      }
    }
  },
  "paths": {
...
    "/api/debts": {
      "get": {
        "summary": "Get all debts",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List of debts" }
        }
      },
      "post": {
        "summary": "Create a debt record",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "person_name": { "type": "string", "example": "Jane Doe" },
                  "direction": { "type": "string", "enum": ["given", "taken"], "example": "given" },
                  "amount": { "type": "number", "example": 250 },
                  "date": { "type": "string", "format": "date-time", "example": "2026-07-21T00:00:00Z" },
                  "due_date": { "type": "string", "format": "date-time", "example": "2026-08-21T00:00:00Z" }
                },
                "required": ["person_name", "direction", "amount", "date"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Debt created successfully" }
        }
      }
    },
    "/api/debts/{id}": {
      "put": {
        "summary": "Update a debt record",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "person_name": { "type": "string" },
                  "direction": { "type": "string", "enum": ["given", "taken"] },
                  "amount": { "type": "number" },
                  "date": { "type": "string", "format": "date-time" },
                  "due_date": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated successfully" }
        }
      },
      "delete": {
        "summary": "Delete a debt record",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Deleted successfully" }
        }
      }
    },
    "/api/debts/{id}/payments": {
      "post": {
        "summary": "Add a payment to a debt",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": { "type": "number", "example": 50 },
                  "date": { "type": "string", "format": "date-time", "example": "2026-07-21T10:00:00Z" },
                  "note": { "type": "string", "example": "First installment" },
                  "transaction_id": { "type": "string" }
                },
                "required": ["amount", "date"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Payment added successfully" }
        }
      }
    },
    "/api/subscriptions": {
      "get": {
        "summary": "Get all subscriptions",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List of subscriptions" }
        }
      },
      "post": {
        "summary": "Create a subscription",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "Netflix" },
                  "amount": { "type": "number", "example": 99.9 },
                  "category_id": { "type": "string", "example": "6a58e3ba24faae1d107d8b8b" },
                  "account_id": { "type": "string", "example": "6a58e5577b7fde809fb38275" },
                  "frequency": { "type": "string", "enum": ["monthly"], "example": "monthly" },
                  "billing_day": { "type": "integer", "minimum": 1, "maximum": 31, "example": 15 },
                  "next_run_date": { "type": "string", "format": "date-time", "example": "2026-08-15T00:00:00Z" },
                  "is_active": { "type": "boolean", "default": true }
                },
                "required": ["name", "amount", "category_id", "account_id", "frequency", "billing_day", "next_run_date"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Subscription created successfully" }
        }
      }
    },
    "/api/subscriptions/{id}": {
      "put": {
        "summary": "Update a subscription",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "amount": { "type": "number" },
                  "category_id": { "type": "string" },
                  "account_id": { "type": "string" },
                  "frequency": { "type": "string", "enum": ["monthly"] },
                  "billing_day": { "type": "integer" },
                  "next_run_date": { "type": "string", "format": "date-time" },
                  "is_active": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated successfully" }
        }
      },
      "delete": {
        "summary": "Delete a subscription",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Deleted successfully" }
        }
      }
    },
...
    "/api/saved-foods": {
      "get": {
        "summary": "Get all saved foods",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List of saved foods" }
        }
      },
      "post": {
        "summary": "Create a saved food",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "food_name": { "type": "string", "example": "Custom Egg Salad" },
                  "serving_description": { "type": "string", "example": "1 porsiyon" },
                  "quantity": { "type": "number", "example": 1 },
                  "fatsecret_food_id": { "type": "string" },
                  "calories": { "type": "number", "example": 220 },
                  "protein_g": { "type": "number", "example": 14 },
                  "carbs_g": { "type": "number", "example": 2 },
                  "fat_g": { "type": "number", "example": 18 }
                },
                "required": ["food_name", "serving_description", "quantity", "calories", "protein_g", "carbs_g", "fat_g"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Saved food created successfully" }
        }
      }
    },
    "/api/saved-foods/{id}": {
      "delete": {
        "summary": "Delete a saved food entry",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Deleted successfully" }
        }
      }
    },
...
  }
}